Reorganise task tests.
