+++ opengever/task/tests/test_forms.py
@@ -1,5 +1,4 @@
 from ftw.testbrowser import browsing
-from opengever.activity import notification_center
 from opengever.activity.model import Activity
 from opengever.activity.model import Subscription
 from opengever.testing import IntegrationTestCase
@@ -7,6 +6,29 @@ from opengever.testing.event_recorder import get_recorded_events
 from opengever.testing.event_recorder import register_event_recorder
 from requests_toolbelt.utils import formdata
 from zope.lifecycleevent.interfaces import IObjectModifiedEvent
+
+
+class TestFormFields(IntegrationTestCase):
+
+    @browsing
+    def test_show_date_of_completion_field_in_edit_form(self, browser):
+        self.login(self.dossier_responsible, browser=browser)
+
+        # seq_subtask_1 is a task with state 'open' and allows editing
+        browser.visit(self.seq_subtask_1, view="edit")
+
+        self.assertNotEqual(
+            'hidden',
+            browser.css('input#form-widgets-date_of_completion').first.type)
+
+    @browsing
+    def test_hide_date_of_completion_field_in_add_form(self, browser):
+        self.login(self.regular_user, browser=browser)
+        browser.open(self.dossier, view='++add++opengever.task.task')
+
+        self.assertEqual(
+            'hidden',
+            browser.css('input#form-widgets-date_of_completion').first.type)
 
 
 class TestTaskAddForm(IntegrationTestCase):
